Refactor exception handling: move classes to utils package and add specific exception handlers
