+++ server/src/main/java/com/kalyvianakis/estiator/io/utils/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.kalyvianakis.estiator.io.global;
+package com.kalyvianakis.estiator.io.utils;
 
 public class ResourceNotFoundException extends Exception {
     public ResourceNotFoundException(String message) {
